requirements.txt
asgiref==3.9.1
coverage==7.10.6
Django==5.2.5
djangorestframework==3.16.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dotenv==1.1.1
sqlparse==0.5.3
